requirements.txt
pika==1.3.0
python-dotenv==0.21.0
requests==2.28.1
retry>=0.9.2
